// frontend-next/src/lib/api/types.ts
export interface components {
    schemas: {
// ...
        TagModel: {
            /** Id */
            id?: number | null;
            /**
             * Createdat
             * Format: date-time
             */
            createdAt?: string;
            /**
             * Updatedat
             * Format: date-time
             */
            updatedAt?: string;
            /**
             * Name
             * @description The name of the tag.
             */
            name: string;
            /**
             * Workspaceid
             * @description The ID of the workspace this tag belongs to.
             */
            workspaceId: number;
            /**
             * Color
             * @description The color of the tag in hex format.
             * @default #E8EAED
             */
            color: string | null;
            /**
             * Userid
             * @description The ID of the user who created the tag.
             */
            userId?: number | null;
        };
// ...
        UnifiedGalleryItemResponse: {
            /** Id */
            id: number;
            /** Workspaceid */
            workspaceId: number;
            /**
             * Createdat
             * Format: date-time
             */
            createdAt: string;
            /** Itemtype */
            itemType: string;
            /** Status */
            status?: string | null;
            /**
             * Gcsuris
             * @default []
             */
            gcsUris: string[];
            /**
             * Thumbnailuris
             * @default []
             */
            thumbnailUris: string[];
            /**
             * Tags
             * @default []
             */
            tags: components["schemas"]["TagModel"][];
            /** Deletedat */
            deletedAt?: string | null;
            /** Workspacename */
            workspaceName?: string | null;
            /** Userpicture */
            userPicture?: string | null;
            /** Useremail */
            userEmail?: string | null;
            /** Metadata */
            metadata?: {
                [key: string]: unknown;
            };
            /**
             * Presignedurls
             * @default []
             */
            presignedUrls: string[];
            /**
             * Presignedthumbnailurls
             * @default []
             */
            presignedThumbnailUrls: string[];
        };
// ...
    };
    responses: never;
    parameters: never;
    requestBodies: never;
    headers: never;
    pathItems: never;
}